Added a scene where you can see all the characters
There is now a scene where you can see all the characters in the database.
Edit and Delete buttons of that scene are not yet implemented.

You can now upload a skill to the database as well.

Fixed a bug with uploading characters if there were no other characters in the database.